Fixed collision bugs and camera stuff
Collision detection for all intents and purposes is fixed.  Now I am
looking to fix the respawn point so that it's a fixed point in the level
and not the camera.  I changed the access to the starting position
variable in the player class to help with player repawn.  Finally
updated the game notes so I know what we are currently working towards